+++ frontend/jest.setup.js
@@ -1,0 +1,3 @@
+// Jest setup file
+require('@testing-library/jest-dom')
+